requirements.txt
tinytag
Pillow
PyPDF2